removed unsued pages + moved log out button
